+++ src/PokemonTabletopAdventures.CoreApi.sln
@@ -9,11 +9,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PokemonTabletopAdventures.C
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PokemonTabletopAdventures.CoreApi.IntegrationTests", "PokemonTabletopAdventures.CoreApi.IntegrationTests\PokemonTabletopAdventures.CoreApi.IntegrationTests.csproj", "{4740268C-C5C0-4E39-8BA7-CBFC08E787E5}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{02EA681E-C7D8-13C7-8484-4AC65E1B71E8}"
-	ProjectSection(SolutionItems) = preProject
-		.editorconfig = .editorconfig
-	EndProjectSection
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
