+++ src/PokemonTabletopAdventures.CoreApi/Domain/GameService.cs
@@ -1,5 +1,4 @@
 ﻿using PokemonTabletopAdventures.CoreApi.Constants;
-using PokemonTabletopAdventures.CoreApi.Domain.Handlers;
 using PokemonTabletopAdventures.CoreApi.DTOs.MongoDB;
 using PokemonTabletopAdventures.CoreApi.Services;
 using PokemonTabletopAdventures.Models.Games;
@@ -11,11 +10,15 @@ public class GameService(
     IRepositoryService repositoryService,
     ITrainerService trainerService,
     INpcService npcService,
-    ISettingService settingService) : AbstractMongoService<GameDto>(repositoryService, MongoCollection.Games), IGameService
+    ISettingService settingService,
+    IDtoToModelMapper dtoToModelMapper,
+    IModelToDtoMapper modelToDtoMapper) : AbstractMongoService<GameDto>(repositoryService, MongoCollection.Games), IGameService
 {
     private readonly ITrainerService _trainerService = trainerService;
     private readonly INpcService _npcService = npcService;
     private readonly ISettingService _settingService = settingService;
+    private readonly IDtoToModelMapper _dtoToModelMapper = dtoToModelMapper;
+    private readonly IModelToDtoMapper _modelToDtoMapper = modelToDtoMapper;
 
     public async Task DeleteGame(Guid id)
     {
@@ -32,7 +35,7 @@ public class GameService(
             name => Collection.GetManyAsync(game => game.Nickname.Contains(name, StringComparison.CurrentCultureIgnoreCase)),
             PropertyNames.Nickname);
 
-        return await Task.WhenAll(dtos.Select(async dto => await DtoHandler.ParseFromDto(dto, false, _npcService, _settingService, _trainerService)));
+        return await Task.WhenAll(dtos.Select(async dto => await _dtoToModelMapper.ParseFromDto(dto, false, _npcService, _settingService, _trainerService)));
     }
 
     public async Task<IEnumerable<Game>> GetAllGamesWithUser(User user)
@@ -42,7 +45,7 @@ public class GameService(
             games => Collection.GetManyAsync(game => games.Contains(game.GameId)),
             PropertyNames.UserGames);
 
-        return await Task.WhenAll(dtos.Select(async dto => await DtoHandler.ParseFromDto(dto, false, _npcService, _settingService, _trainerService)));
+        return await Task.WhenAll(dtos.Select(async dto => await _dtoToModelMapper.ParseFromDto(dto, false, _npcService, _settingService, _trainerService)));
     }
 
     public async Task<Game> GetGame(Guid id, bool isGM)
@@ -52,7 +55,7 @@ public class GameService(
             id => Collection.GetOneAsync(game => game.GameId == id),
             PropertyNames.GameId);
 
-        return await DtoHandler.ParseFromDto(dto, isGM, _npcService, _settingService, _trainerService);
+        return await _dtoToModelMapper.ParseFromDto(dto, isGM, _npcService, _settingService, _trainerService);
     }
 
     public async Task<string> GetGameNickname(Guid gameId)
@@ -68,7 +71,7 @@ public class GameService(
             games => Collection.GetManyAsync(x => !games.Contains(x.GameId), 0, 20),
             PropertyNames.GameId);
 
-        return await Task.WhenAll(dtos.Select(async dto => await DtoHandler.ParseFromDto(dto, false, _npcService, _settingService, _trainerService)));
+        return await Task.WhenAll(dtos.Select(async dto => await _dtoToModelMapper.ParseFromDto(dto, false, _npcService, _settingService, _trainerService)));
     }
 
     public async Task<bool> HasGM(Guid gameId)
@@ -80,7 +83,7 @@ public class GameService(
 
     public async Task PostGame(Game game, string passwordHash)
     {
-        var dto = DtoHandler.ParseFromModel(game);
+        var dto = await _modelToDtoMapper.ParseFromModel(game);
         dto.PasswordHash = passwordHash;
         await PostDocument(dto);
     }
@@ -92,7 +95,7 @@ public class GameService(
             game => game.GameId == theGame.GameId,
             new Models.UpdateData(PropertyNames.Logs, theGame.Logs?.Union(logs) ?? logs));
 
-        return await DtoHandler.ParseFromDto(dto, isGM, _npcService, _settingService, _trainerService);
+        return await _dtoToModelMapper.ParseFromDto(dto, isGM, _npcService, _settingService, _trainerService);
     }
 
     public async Task<Game> UpdateGameNpcList(Guid gameId, IEnumerable<Guid> npcIds)
@@ -102,7 +105,7 @@ public class GameService(
             game => game.GameId == gameId,
             new Models.UpdateData(PropertyNames.Npcs, npcIds));
 
-        return await DtoHandler.ParseFromDto(dto, true, _npcService, _settingService, _trainerService);
+        return await _dtoToModelMapper.ParseFromDto(dto, true, _npcService, _settingService, _trainerService);
     }
 
     public async Task<Game> UpdateGameOnlineStatus(Guid gameId, bool isOnline)
@@ -112,6 +115,6 @@ public class GameService(
             game => game.GameId == gameId,
             new Models.UpdateData(PropertyNames.IsOnline, isOnline));
 
-        return await DtoHandler.ParseFromDto(dto, true, _npcService, _settingService, _trainerService);
+        return await _dtoToModelMapper.ParseFromDto(dto, true, _npcService, _settingService, _trainerService);
     }
 }
