+++ src/PokemonTabletopAdventures.CoreApi/Domain/PokedexService.cs
@@ -1,5 +1,4 @@
 ﻿using PokemonTabletopAdventures.CoreApi.Constants;
-using PokemonTabletopAdventures.CoreApi.Domain.Handlers;
 using PokemonTabletopAdventures.CoreApi.DTOs.MongoDB;
 using PokemonTabletopAdventures.CoreApi.Exceptions;
 using PokemonTabletopAdventures.CoreApi.Services;
@@ -7,8 +6,11 @@ using PokemonTabletopAdventures.Models.Pokedex;
 
 namespace PokemonTabletopAdventures.CoreApi.Domain;
 
-public class PokedexService(IRepositoryService repositoryService) : AbstractMongoService<PokeDexItemDto>(repositoryService, MongoCollection.Pokedex), IPokedexService
+public class PokedexService(
+    IRepositoryService repositoryService,
+    IDtoToModelMapper dtoToModelMapper) : AbstractMongoService<PokeDexItemDto>(repositoryService, MongoCollection.Pokedex), IPokedexService
 {
+    private readonly IDtoToModelMapper _dtoToModelMapper = dtoToModelMapper;
     public async Task DeleteDexItemForTrainer(Guid trainerId, Guid gameId)
     {
         await ThrowIfNull(
@@ -22,7 +24,7 @@ public class PokedexService(IRepositoryService repositoryService) : AbstractMong
         var dto = await Collection.GetOneAsync(dexItem => dexItem.TrainerId == trainerId && dexItem.GameId == gameId && dexItem.DexNo == dexNo)
              ?? throw new UnknownEntityException<PokedexItem>(PropertyNames.DexNo, dexNo);
 
-        return DtoHandler.ParseFromDto(dto);
+        return await _dtoToModelMapper.ParseFromDto(dto);
     }
 
     public async Task<IEnumerable<PokedexItem>> GetTrainerPokeDex(Guid trainerId, Guid gameId)
@@ -32,7 +34,7 @@ public class PokedexService(IRepositoryService repositoryService) : AbstractMong
             x => Collection.GetManyAsync(dexItem => dexItem.TrainerId == x.trainerId && dexItem.GameId == x.gameId),
             $"{PropertyNames.TrainerId} {PropertyNames.GameId}");
 
-        return dtos.Select(DtoHandler.ParseFromDto);
+        return await Task.WhenAll(dtos.Select(_dtoToModelMapper.ParseFromDto));
     }
 
     public async Task PostDexItem(Guid trainerId, Guid gameId, int dexNo, bool isSeen, bool isCaught)
@@ -56,7 +58,7 @@ public class PokedexService(IRepositoryService repositoryService) : AbstractMong
             dexItem => dexItem.TrainerId == trainerId && dexItem.GameId == gameId && dexItem.DexNo == dexNo,
             new Models.UpdateData(PropertyNames.IsSeen, true));
 
-        return DtoHandler.ParseFromDto(dto);
+        return await _dtoToModelMapper.ParseFromDto(dto);
     }
 
     public async Task<PokedexItem> UpdateDexItemIsSeen(Guid trainerId, Guid gameId, int dexNo)
@@ -67,6 +69,6 @@ public class PokedexService(IRepositoryService repositoryService) : AbstractMong
             new Models.UpdateData(PropertyNames.IsSeen, true),
             new Models.UpdateData(PropertyNames.IsCaught, true));
 
-        return DtoHandler.ParseFromDto(dto);
+        return await _dtoToModelMapper.ParseFromDto(dto);
     }
 }
