+++ src/PokemonTabletopAdventures.CoreApi/Domain/RepositoryService.cs
@@ -2,7 +2,6 @@
 using MongoDB.Bson.Serialization.Serializers;
 using MongoDB.Driver;
 using PokemonTabletopAdventures.CoreApi.Constants;
-using PokemonTabletopAdventures.CoreApi.DTOs;
 using PokemonTabletopAdventures.CoreApi.Services;
 using System.Diagnostics.CodeAnalysis;
