+++ src/PokemonTabletopAdventures.CoreApi/Domain/ShopService.cs
@@ -1,5 +1,4 @@
 ﻿using PokemonTabletopAdventures.CoreApi.Constants;
-using PokemonTabletopAdventures.CoreApi.Domain.Handlers;
 using PokemonTabletopAdventures.CoreApi.DTOs.MongoDB;
 using PokemonTabletopAdventures.CoreApi.Services;
 using PokemonTabletopAdventures.Models.Settings;
@@ -7,8 +6,14 @@ using PokemonTabletopAdventures.Models.Shops;
 
 namespace PokemonTabletopAdventures.CoreApi.Domain;
 
-public class ShopService(IRepositoryService repositoryService) : AbstractMongoService<ShopDto>(repositoryService, MongoCollection.Shops), IShopService
+public class ShopService(
+    IRepositoryService repositoryService,
+    IDtoToModelMapper dtoToModelMapper,
+    IModelToDtoMapper modelToDtoMapper) : AbstractMongoService<ShopDto>(repositoryService, MongoCollection.Shops), IShopService
 {
+    private readonly IDtoToModelMapper _dtoToModelMapper = dtoToModelMapper;
+    private readonly IModelToDtoMapper _modelToDtoMapper = modelToDtoMapper;
+
     public async Task DeleteShop(Guid id, Guid gameId)
     {
         await ThrowIfNull(
@@ -32,7 +37,7 @@ public class ShopService(IRepositoryService repositoryService) : AbstractMongoSe
             id => Collection.GetManyAsync(shop => shop.GameId == id),
             PropertyNames.GameId);
 
-        return dtos.Select(DtoHandler.ParseFromDto);
+        return await Task.WhenAll(dtos.Select(_dtoToModelMapper.ParseFromDto));
     }
 
     public async Task<Shop> GetShopById(Guid id, Guid gameId)
@@ -42,7 +47,7 @@ public class ShopService(IRepositoryService repositoryService) : AbstractMongoSe
             id => Collection.GetOneAsync(shop => shop.GameId == gameId && shop.ShopId == id),
             PropertyNames.ShopId);
 
-        return DtoHandler.ParseFromDto(dto);
+        return await _dtoToModelMapper.ParseFromDto(dto);
     }
 
     public async Task<IEnumerable<Shop>> GetShopsBySetting(Setting setting)
@@ -53,18 +58,18 @@ public class ShopService(IRepositoryService repositoryService) : AbstractMongoSe
             id => Collection.GetManyAsync(shop => shopIds.Contains(shop.ShopId) && setting.GameId == shop.GameId),
             PropertyNames.SettingShops);
 
-        return dtos.Select(DtoHandler.ParseFromDto);
+        return await Task.WhenAll(dtos.Select(_dtoToModelMapper.ParseFromDto));
     }
 
     public async Task PostShop(Shop shop)
     {
-        var dto = DtoHandler.ParseFromModel(shop);
+        var dto = await _modelToDtoMapper.ParseFromModel(shop);
         await PostDocument(dto);
     }
 
     public async Task<Shop> UpdateShop(Shop updatedShop)
     {
-        var dto = DtoHandler.ParseFromModel(updatedShop);
+        var dto = await _modelToDtoMapper.ParseFromModel(updatedShop);
         await UpsertDocument(
             shop => shop.ShopId,
             updatedShop.ShopId,
