+++ src/PokemonTabletopAdventures.CoreApi/Domain/UserMessageThreadService.cs
@@ -1,13 +1,18 @@
 ﻿using PokemonTabletopAdventures.CoreApi.Constants;
-using PokemonTabletopAdventures.CoreApi.Domain.Handlers;
 using PokemonTabletopAdventures.CoreApi.DTOs.MongoDB;
 using PokemonTabletopAdventures.CoreApi.Services;
 using PokemonTabletopAdventures.Models.Users;
 
 namespace PokemonTabletopAdventures.CoreApi.Domain;
 
-public class UserMessageThreadService(IRepositoryService repositoryService) : AbstractMongoService<UserMessageThreadDto>(repositoryService, MongoCollection.UserMessageThreads), IUserMessageThreadService
+public class UserMessageThreadService(
+    IRepositoryService repositoryService,
+    IDtoToModelMapper dtoToModelMapper,
+    IModelToDtoMapper modelToDtoMapper) : AbstractMongoService<UserMessageThreadDto>(repositoryService, MongoCollection.UserMessageThreads), IUserMessageThreadService
 {
+    private readonly IDtoToModelMapper _dtoToModelMapper = dtoToModelMapper;
+    private readonly IModelToDtoMapper _modelToDtoMapper = modelToDtoMapper;
+
     public async Task<UserMessageThread> GetMessageById(Guid id)
     {
         var dto = await ThrowIfNull(
@@ -15,18 +20,18 @@ public class UserMessageThreadService(IRepositoryService repositoryService) : Ab
             id => Collection.GetOneAsync(message => message.MessageId == id),
             PropertyNames.MessageId);
 
-        return DtoHandler.ParseFromDto(dto);
+        return await _dtoToModelMapper.ParseFromDto(dto);
     }
 
     public async Task PostThread(UserMessageThread thread)
     {
-        var dto = DtoHandler.ParseFromModel(thread);
+        var dto = await _modelToDtoMapper.ParseFromModel(thread);
         await PostDocument(dto);
     }
 
     public async Task<UserMessageThread> UpdateThread(UserMessageThread updatedThread)
     {
-        var dto = DtoHandler.ParseFromModel(updatedThread);
+        var dto = await _modelToDtoMapper.ParseFromModel(updatedThread);
         await UpsertDocument(
             thread => thread.MessageId,
             updatedThread.MessageId,
